use boost::program_options to parse the command line
this allows to use standard long and short options (-o and --option) and
should make easier the addition of new options in the future.
Input are now read from stdin when no files are passed, as do most of unix
CLI programs.
The old CLI is kept for backward compatibility, but not shown in the help
message.