report options now take the output file name as argument - closes #20
There are now 3 options to control the output format of vera: --std-report,
--vc-report and --xml-report. This options take a file as argument and may be
used several time. By default and unless --quiet is used, --std-report write
its output to "-" (the standard or error output). The other report are not
written by default.
This allows for example to write the xml report in a file and the vc report
to the standard output with a command like:

  vera++ --vc-report=- --xml-report=/path/to/report.xml file.cpp file2.cpp