use the wiki to generate the groff manpage and the html documentation
The doc is generated by converting the mardown format from the wiki with pandoc

  http://johnmacfarlane.net/pandoc/

The doc is not generated if pandoc is not found by cmake, or if the wiki source
are not found. A warning is generated at configuration time if the doc can't be
generated.
A markdown manpage header makes the generated manpage match a bit better the
usual manpage format.