enhanced exclusion mechanism - see #23
the exclusion files are now parsed as a set of regular expressions that match
the log output in the standard format with the rule name. If the regular
expression match, the report is excluded.
For example, to exclude all the T003 report for the 'case' and  'delete'
keyword, just add this line to ane exclusion file

    :\d+: T003: keyword '(case|delete)' not followed by a single space

the empty lines are ignored in the exclusion files as well as the lines beginning
with a #, so the file can be commented and formatted a bit